+++ crates/teloxide-core/src/types/update.rs
@@ -613,6 +613,7 @@ mod test {
                     }),
                     reply_markup: None,
                     author_signature: None,
+                    effect_id: None,
                     is_automatic_forward: false,
                     has_protected_content: false,
                     is_from_offline: false,